QString/doc: correct the record on const char* optimizations
This portion of the documentation was there since the Qt 4.5 import of
the repository and may have been correct at the time. They haven't been
for some time and definitely aren't now. So be clear that even if there
are overloads, some of them are bad ideas.

Change-Id: I810d70e579eb4e2c8e45fffd1719adefb6f9f3bf
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit d46eeffe83db56aea7703d20351dde2969bf372a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>