Bootstrap: explicitly include stdlib.h to check for __GLIBC_PREREQ
Previously the code was relying on the includes in qglobal.h, but now
when we try to get rid of qglobal.h we can end up in a situation when
qconfig-bootstrap.h was first included before qglobal.h (__GLIBC_PREREQ
not defined), and also included once again after qglobal.h
(__GLIBC_PREREQ defined, at least on linux). This resulted in
redefenition of Qt features dependent on __GLIBC_PREREQ.

This patch fixes it by explicitly including the stdlib.h header which
will provide __GLIBC_PREREQ definition when it is available.

Task-number: QTBUG-107046
Change-Id: Idbf7a11151c5f81723131daf25c06ef454ff5cbb
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>